Final_Basys3.ver/motion_game_ver2/requirements.txt
pygame
pyserial
opencv-python
numpy
pillow
